# Import rasterio within the `rasterio_crs` property to speedup startup