--- a/morecantile/__init__.py
+++ b/morecantile/__init__.py
@@ -8,7 +8,7 @@ Refs:
 
 """
 
-__version__ = "3.0.0a0"
+__version__ = "3.0.0a1"
 
 from .commons import BoundingBox, Coords, Tile  # noqa
 from .defaults import tms  # noqa