--- a/morecantile/models.py
+++ b/morecantile/models.py
@@ -1,4 +1,5 @@
 """Pydantic modules for OGC TileMatrixSets (https://www.ogc.org/standards/tms)"""
+
 import math
 import warnings
 from typing import Any, Dict, Iterator, List, Optional, Sequence, Tuple, Union