--- a/setup.cfg
+++ b/setup.cfg
@@ -1,6 +1,6 @@
 
 [bumpversion]
-current_version = 3.0.0a0
+current_version = 3.0.0a1
 commit = True
 tag = True
 tag_name = {new_version}