--- a/setup.py
+++ b/setup.py
@@ -16,7 +16,7 @@ extra_reqs = {
 
 setup(
     name="morecantile",
-    version="3.0.0a0",
+    version="3.0.0a1",
     python_requires=">=3.7",
     description=u"""Construct and use map tile grids (a.k.a TileMatrixSet / TMS).""",
     long_description=long_description,