ref https://github.com/developmentseed/morecantile/issues/64#issuecomment-921648142

as mentioned in ☝️ Importing rasterio if not needed will slow down the startup. 

### Before 
```python
%time import morecantile
CPU times: user 258 ms, sys: 35.9 ms, total: 294 ms
Wall time: 360 ms

tms = morecantile.tms.get("WebMercatorQuad")

%time tms.rasterio_crs
CPU times: user 3.11 ms, sys: 479 µs, total: 3.59 ms
Wall time: 3.59 ms
```

### With this PR
```python
%time import morecantile
CPU times: user 102 ms, sys: 12.5 ms, total: 114 ms
Wall time: 114 ms

tms = morecantile.tms.get("WebMercatorQuad")

%time tms.rasterio_crs
CPU times: user 158 ms, sys: 33.6 ms, total: 192 ms
Wall time: 258 ms

# makes sure nothing happens after
%time tms.rasterio_crs
CPU times: user 108 µs, sys: 16 µs, total: 124 µs
Wall time: 125 µs
```

in ☝️ we can see that importing rasterio within the `rasterio_crs` only slows down the first call to the properties in addition of speeding up the general morecantile import